requirements.txt
djangorestframework
PyPDF2
python-docx
nltk
scikit-learn
spacy
